+++ src/components/ActionMenu.vue
@@ -1,5 +1,5 @@
 <template>
-  <popover :active="isActive" :align="alignLeft" bottom>
+  <popover :visible="visible" :align="alignLeft" bottom @closed="close">
     <div class="action-menu__body">
       <div class="action-menu__section">
         <div class="action-menu__option" @click="createStep('duplicate')">Duplicate column</div>
@@ -15,7 +15,7 @@
   </popover>
 </template>
 <script lang="ts">
-import { Component, Prop, Vue, Watch } from 'vue-property-decorator';
+import { Component, Prop, Vue } from 'vue-property-decorator';
 
 import { POPOVER_ALIGN } from '@/components/constants';
 import { Pipeline, PipelineStep, PipelineStepName } from '@/lib/steps';
@@ -32,16 +32,16 @@ import Popover from './Popover.vue';
 })
 export default class ActionMenu extends Vue {
   @Prop({
-    type: Boolean,
-    default: () => false,
-  })
-  isActive!: boolean;
-
-  @Prop({
     type: String,
     default: () => '',
   })
   columnName!: string;
+
+  @Prop({
+    type: Boolean,
+    default: true,
+  })
+  visible!: boolean;
 
   @VQBModule.Getter computedActiveStepIndex!: number;
   @VQBModule.Getter isEditingStep!: boolean;
@@ -52,18 +52,6 @@ export default class ActionMenu extends Vue {
   @VQBModule.Mutation closeStepForm!: () => void;
 
   alignLeft: string = POPOVER_ALIGN.LEFT;
-
-  /**
-   * @description Close the popover when clicking outside
-   */
-  clickListener(e: Event) {
-    const target = e.target as HTMLElement;
-    const hasClickOnItSelf = target === this.$el || this.$el.contains(target);
-
-    if (!hasClickOnItSelf) {
-      this.close();
-    }
-  }
 
   close() {
     this.$emit('closed');
@@ -106,15 +94,6 @@ export default class ActionMenu extends Vue {
     this.setPipeline({ pipeline: newPipeline });
     this.selectStep({ index });
     this.close();
-  }
-
-  @Watch('isActive')
-  onIsActiveChanged(val: boolean) {
-    if (val) {
-      window.addEventListener('click', this.clickListener, { capture: true });
-    } else {
-      window.removeEventListener('click', this.clickListener, { capture: true });
-    }
   }
 }
 </script>